fix(eslint): fix all eslint errors

+++ controllers/clothingItems.js
@@ -46,6 +46,13 @@ const deleteItem = async (req, res) => {
       return;
     }
 
+    const itemToDelete = await Item.findById(itemId).orFail();
+    if (!itemToDelete.owner === req.user._id) {
+      res.status(STATUS_CODES.FORBIDDEN)
+      .json({message: "User does not own clothing item"})
+      return
+    }
+
     const deletedItem = await Item.findByIdAndDelete(itemId).orFail();
     res
       .status(200)
